requirements.txt
absl-py==1.4.0
ansi2html==1.8.0
certifi==2023.7.22
charset-normalizer==3.2.0
chex==0.1.82
click==8.1.6
dash==2.11.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dm-haiku==0.0.10
Flask==2.2.5
idna==3.4
itsdangerous==2.1.2
jax==0.4.13
jaxlib==0.4.13+cuda12.cudnn89
Jinja2==3.1.2
jmp==0.0.4
MarkupSafe==2.1.3
ml-dtypes==0.2.0
nest-asyncio==1.5.6
numpy==1.25.1
nvidia-cublas-cu12==12.2.1.16
nvidia-cuda-cupti-cu12==12.2.60
nvidia-cuda-nvcc-cu12==12.2.91
nvidia-cuda-runtime-cu12==12.2.53
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.8.15
nvidia-cusolver-cu12==11.5.0.53
nvidia-cusparse-cu12==12.1.1.53
nvidia-nvjitlink-cu12==12.2.91
opt-einsum==3.3.0
optax==0.1.5
packaging==23.1
pandas==2.0.3
plotly==5.15.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
retrying==1.3.4
scipy==1.11.1
six==1.16.0
tabulate==0.9.0
tenacity==8.2.2
toolz==0.12.0
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
Werkzeug==2.2.3
